split out fits utils into a separate file. Don't use load_fits/save_fits in mosaic app
