+++ spimple/utils.py
@@ -15,9 +15,8 @@ import xarray as xr
 from scipy.interpolate import RegularGridInterpolator
 from scipy import ndimage
 import ray
-from astropy.time import Time
-from casacore.quanta import quantity
-from datetime import datetime, timezone
+from time import time
+from spimple.fits import save_fits, load_fits, set_wcs, data_from_header
 
 def str2bool(v):
     """
@@ -35,204 +34,6 @@ def str2bool(v):
     else:
         import argparse
         raise argparse.ArgumentTypeError('Boolean value expected.')
-
-def to4d(data):
-    if data.ndim == 4:
-        return data
-    elif data.ndim == 2:
-        return data[None, None]
-    elif data.ndim == 3:
-        return data[None]
-    elif data.ndim == 1:
-        return data[None, None, None]
-    else:
-        raise ValueError("Only arrays with ndim <= 4 can be broadcast to 4D.")
-
-
-def data_from_header(hdr, axis=3):
-    npix = hdr['NAXIS' + str(axis)]
-    refpix = hdr['CRPIX' + str(axis)]
-    delta = hdr['CDELT' + str(axis)]
-    ref_val = hdr['CRVAL' + str(axis)]
-    return ref_val + np.arange(1 - refpix, 1 + npix - refpix) * delta, ref_val
-
-
-def load_fits(name, dtype=np.float32):
-    data = fits.getdata(name)
-    data = np.transpose(to4d(data), axes=(1, 0, 3, 2))  # fits and beams table
-    return np.require(data, dtype=dtype, requirements='C')
-
-
-def save_fits(data, name, hdr, overwrite=True, dtype=np.float32, beams_hdu=None):
-    hdu = fits.PrimaryHDU(header=hdr)
-    data = np.transpose(to4d(data), axes=(1, 0, 3, 2))
-    hdu.data = np.require(data, dtype=dtype, requirements='F')
-    if beams_hdu is not None:
-        hdul = fits.HDUList([hdu, beams_hdu])
-        hdul.writeto(name, overwrite=overwrite)
-    else:
-        hdu.writeto(name, overwrite=overwrite)
-    return
-
-
-def set_wcs(cell_x, cell_y, nx, ny, radec, freq,
-            unit='Jy/beam', GuassPar=None, ms_time=None,
-            header=True, casambm=True):
-    """
-    cell_x/y - cell sizes in degrees
-    nx/y - number of x and y pixels
-    radec - right ascention and declination in radians
-    freq - frequencies in Hz
-    unit - Jy/beam or Jy/pixel
-    GuassPar - MFS beam parameters in degrees
-    ms_time - measurement set time
-    header - if True, return a header, otherwise return a WCS object
-    casambm - if True, add the CASAMBM keyword to the header
-    """
-
-    w = WCS(naxis=4)
-    w.wcs.ctype = ['RA---SIN', 'DEC--SIN', 'FREQ', 'STOKES']
-    w.wcs.cdelt[0] = -cell_x
-    w.wcs.cdelt[1] = cell_y
-    w.wcs.cdelt[3] = 1
-    w.wcs.cunit[0] = 'deg'
-    w.wcs.cunit[1] = 'deg'
-    w.wcs.cunit[2] = 'Hz'
-    w.wcs.cunit[3] = ''
-    if np.size(freq) > 1:
-        nchan = freq.size
-        crpix3 = nchan//2+1
-        ref_freq = freq[crpix3]
-        df = freq[1]-freq[0]
-        w.wcs.cdelt[2] = df
-    else:
-        if isinstance(freq, np.ndarray) and freq.size == 1:
-            ref_freq = freq[0]
-        else:
-            ref_freq = freq
-        crpix3 = 1
-    w.wcs.crval = [radec[0]*180.0/np.pi, radec[1]*180.0/np.pi, ref_freq, 1]
-    w.wcs.crpix = [1 + nx//2, 1 + ny//2, crpix3, 1]
-    w.wcs.equinox = 2000.0
-
-    if header:
-        header = w.to_header()
-        header['RESTFRQ'] = ref_freq
-        header['ORIGIN'] = 'pfb-imaging'
-        header['BTYPE'] = 'Intensity'
-        header['BUNIT'] = unit
-        header['SPECSYS'] = 'TOPOCENT'
-        if ms_time is not None:
-            # TODO - probably a round about way of doing this
-            unix_time = quantity(f'{ms_time}s').to_unix_time()
-            utc_iso = datetime.fromtimestamp(unix_time,
-                            tz=timezone.utc).strftime('%Y-%m-%d %H:%M:%S')
-            header['UTC_TIME'] = utc_iso
-            t = Time(utc_iso)
-            t.format = 'fits'
-            header['DATE-OBS'] = t.value
-
-        # What are these used for?
-        # if 'LONPOLE' in header:
-        #     header.pop('LONPOLE')
-        # if 'LATPOLE' in header:
-        #     header.pop('LATPOLE')
-        # if 'RADESYS' in header:
-        #     header.pop('RADESYS')
-        # if 'MJDREF' in header:
-        #     header.pop('MJDREF')
-
-        # header['EQUINOX'] = 2000.0
-        header['BSCALE'] = 1.0
-        header['BZERO'] = 0.0
-        if casambm:
-            header['CASAMBM'] = casambm  # we need this to pick up the beams table
-        
-        if GuassPar is not None:
-            header = add_beampars(header, GuassPar)
-
-        return header
-    else:
-        return w
-
-
-def add_beampars(hdr, GaussPar, GaussPars=None, unit2deg=1.0):
-    """
-    Add beam keywords to header.
-    GaussPar - MFS beam pars
-    GaussPars - beam pars for cube
-    unit2deg - conversion factor to convert BMAJ/BMIN to degrees
-
-    PA is passed in radians and follows the parametrisation in
-    
-    pfb/utils/misc/Gaussian2D
-    
-    """
-    if len(GaussPar) == 1:
-        GaussPar = GaussPar[0]
-    elif len(GaussPar) != 3:
-        raise ValueError('Invalid value for GaussPar')
-
-    if not np.isnan(GaussPar).any():
-        hdr['BMAJ'] = GaussPar[0]*unit2deg
-        hdr['BMIN'] = GaussPar[1]*unit2deg
-        hdr['BPA'] = GaussPar[2]*180/np.pi
-
-    if GaussPars is not None:
-        for i in range(len(GaussPars)):
-            if not np.isnan(GaussPars[i]).any():
-                hdr['BMAJ' + str(i+1)] = GaussPars[i][0]*unit2deg
-                hdr['BMIN' + str(i+1)] = GaussPars[i][1]*unit2deg
-                hdr['BPA' + str(i+1)] = GaussPars[i][2]*180/np.pi
-
-    return hdr
-
-
-def set_header_info(mhdr, ref_freq, freq_axis, beampars=None):
-    """
-    Creates a new FITS header with updated frequency axis and 
-    optional beam parameters.
-    
-    Copies selected header keys from the input header, sets the
-    specified frequency axis to length 1 with the given reference
-    frequency, and optionally adds beam parameters (`BMAJ`, `BMIN`, `BPA`)
-    if provided.
-    
-    Args:
-        mhdr: Input FITS header to copy keys from.
-        ref_freq: Reference frequency value to set on the specified axis.
-        freq_axis: Axis index (3 or 4) to update with the reference frequency.
-        beampars: Optional tuple of (major axis, minor axis, position angle)
-                  for beam parameters.
-    
-    Returns:
-        A new astropy.io.fits.Header object with updated frequency and
-        optional beam information.
-    """
-    hdr_keys = ['SIMPLE', 'BITPIX', 'NAXIS', 'NAXIS1', 'NAXIS2', 'NAXIS3',
-                'NAXIS4', 'CTYPE1', 'CTYPE2', 'CTYPE3', 'CTYPE4', 'CRPIX1',
-                'CRPIX2', 'CRPIX3', 'CRPIX4', 'CRVAL1', 'CRVAL2', 'CRVAL3',
-                'CRVAL4', 'CDELT1', 'CDELT2', 'CDELT3', 'CDELT4']
-
-    new_hdr = {}
-    for key in hdr_keys:
-        new_hdr[key] = mhdr[key]
-
-    if freq_axis == 3:
-        new_hdr["NAXIS3"] = 1
-        new_hdr["CRVAL3"] = ref_freq
-    elif freq_axis == 4:
-        new_hdr["NAXIS4"] = 1
-        new_hdr["CRVAL4"] = ref_freq
-
-    if beampars is not None:
-        new_hdr['BMAJ'] = beampars[0]
-        new_hdr['BMIN'] = beampars[1]
-        new_hdr['BPA'] = beampars[2]
-
-    new_hdr = fits.Header(new_hdr)
-
-    return new_hdr
 
 
 def Gaussian2D(xin, yin, GaussPar=(1., 1., 0.), normalise=True, nsigma=5):
@@ -606,7 +407,7 @@ def mosaic_info(im_list, oname, ref_image=None):
     freqs = []
     flatfreqs = []
     shapes = []
-    basename = oname.rstrip('.fits')
+    basename = oname.removesuffix('.fits')
     for imnum, im in enumerate(im_list):
         hdr = fits.getheader(im)
         nu = data_from_header(hdr, axis=3)[0]
@@ -626,10 +427,7 @@ def mosaic_info(im_list, oname, ref_image=None):
     ref_wcs, shape_out = find_optimal_celestial_wcs(
                                     list(zip(shapes, wcss)),
                                     projection='SIN')
-    # ref_hdr = ref_wcs.to_header()
-    # ref_hdr['NAXIS1'] = shape_out[0]
-    # ref_hdr['NAXIS2'] = shape_out[1]
-    ref_wcs.array_shape = (shape_out[1], shape_out[0])
+    ref_wcs.array_shape = (shape_out[0], shape_out[1])
 
     return ref_wcs, ufreqs, out_names
 
@@ -656,7 +454,8 @@ def project(im, imnum, ref_wcs, beam, oname, band='L', method='interp'):
     beamo = RegularGridInterpolator(
                             (bfreq, l_beam, m_beam),
                             pbeam,
-                            bounds_error=True,
+                            bounds_error=False,
+                            fill_value=None,
                             method='linear')
     
     cell_x, cell_y = ref_wcs.wcs.cdelt
@@ -670,7 +469,6 @@ def project(im, imnum, ref_wcs, beam, oname, band='L', method='interp'):
     # image = load_fits(im)
     image = fits.getdata(im)
     ncorr, nchan, nx, ny = image.shape
-    print('Loading header for ', im)
     hdr = fits.getheader(im)
     freq, _ = data_from_header(hdr, axis=3)
     wcs = WCS(hdr).dropaxis(-1).dropaxis(-1)
@@ -678,35 +476,45 @@ def project(im, imnum, ref_wcs, beam, oname, band='L', method='interp'):
     nx, ny = wcs.array_shape
     l = (-(nx)//2 + np.arange(nx)) * cxi
     m = (-(ny)//2 + np.arange(ny)) * cyi
-    ll, mm = np.meshgrid(l, m)
+    ll, mm = np.meshgrid(l, m, indexing='ij')
     
-    basename = oname.rstrip('.fits')
+    basename = oname.removesuffix('.fits')
     # TODO - make these in parallel
     for c in range(ncorr):
         for f in range(nchan):
             bdata = np.zeros((nx, ny), dtype=np.float64)
+            # ti = time()
             beami = beamo((freq[c], ll, mm))
+            # print('Interp ', time() - ti)
+            # ti = time()
             step = 25
             angles = np.linspace(0, 359, step)
             for angle in angles:
-                bdata += ndimage.rotate(beami, angle, reshape=False, mode='nearest')
+                bdata += ndimage.rotate(beami, angle,
+                                        reshape=False,
+                                        order=1,
+                                        mode='nearest')
             bdata /= angles.size
-
+            # print('azimuthal average ', time() - ti)
             # project beam
+            # ti = time()
             pbeam, footprint = reproject_interp((bdata, wcs),
                                                 ref_wcs,
                                                 shape_out=(nxo, nyo),
-                                                blocksize='auto',
+                                                block_size='auto',
                                                 parallel=4)
             mask = footprint > 0
             pbeam[~mask] = 0
+            # print('beam project ', time() - ti)
             
             # project image
+            # ti = time()
             pdata, _ = reproject_interp((image[c, f], wcs),
                                         ref_wcs,
                                         shape_out=(nxo, nyo),
-                                        blocksize='auto',
+                                        block_size='auto',
                                         parallel=4)
+            # print('Image project ', time() - ti)
 
             im_attrs = {
             'freq': freq[c],
@@ -727,10 +535,25 @@ def project(im, imnum, ref_wcs, beam, oname, band='L', method='interp'):
                        compute=True, mode='w',
                        consolidated=False)
             
-            # fname = ds_name.rstrip('.zarr') + '.fits'
-            # import ipdb; ipdb.set_trace()
-            # save_fits(pdata, fname, ref_wcs)
             
+            # hdri = wcs.to_header()
+            # hdu = fits.PrimaryHDU(header=hdri)
+            # hdu.data = image[c, f]
+            # fname = ds_name.removesuffix('.zarr') + '_image.fits'
+            # hdu.writeto(fname, overwrite=True)
+            # hdu.data = bdata
+            # fname = ds_name.removesuffix('.zarr') + '_beam.fits'
+            # hdu.writeto(fname, overwrite=True)
+
+            # ref_hdr = ref_wcs.to_header()
+            # hdu = fits.PrimaryHDU(header=ref_hdr)
+            # hdu.data = pdata
+            # fname = ds_name.removesuffix('.zarr') + '_pimage.fits'
+            # hdu.writeto(fname, overwrite=True)
+            # hdu.data = pbeam
+            # fname = ds_name.removesuffix('.zarr') + '_pbeam.fits'
+            # hdu.writeto(fname, overwrite=True)
+
     return imnum
             
 
@@ -790,8 +613,7 @@ def stitch_images(freq, im_list, eta=1e-3):
 #     nx, ny = wcs.array_shape
 #     x_pix, y_pix = np.mgrid[0:nx, 0:ny]
 #     ra, dec = wcs.pixel_to_world(x_pix, y_pix)
-    
-    
+
